chore: Remove self-install telemetry
Converted the `Telemetry` component into `useTelemetryInitializer` composable
since it no longer has any DOM elements.

See Linear issue for reasoning behind this change.
